+++ src/main/java/at/hannibal2/skyhanni/config/features/Mobs.java
@@ -66,4 +66,9 @@ public class Mobs {
     @ConfigEditorBoolean
     @ConfigAccordionId(id = 1)
     public boolean areaBossRespawnTimer = false;
+
+    @Expose
+    @ConfigOption(name = "Enderman Teleportation Hider", desc = "Stops the Enderman Teleportation animation (Like in SBA)")
+    @ConfigEditorBoolean
+    public boolean endermanTeleportationHider = false;
 }